Remove the distinction between Async and Sync protocol requests.
See also hazelcast/hazelcast#6711
