+++ Hazelcast.Net/Hazelcast.Net.csproj
@@ -26,6 +26,8 @@
     <Prefer32Bit>false</Prefer32Bit>
     <DocumentationFile>bin\Debug\Hazelcast.Net.xml</DocumentationFile>
     <CodeAnalysisRuleSet>MinimumRecommendedRules.ruleset</CodeAnalysisRuleSet>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+    <NoWarn>1591</NoWarn>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>pdbonly</DebugType>
@@ -58,13 +60,12 @@
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapAddInterceptorCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapEvictAllCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapRemoveInterceptorCodec.cs" />
-    <Compile Include="Hazelcast.Client.Protocol.Codec\RingbufferAddAllAsyncCodec.cs" />
-    <Compile Include="Hazelcast.Client.Protocol.Codec\RingbufferAddAsyncCodec.cs" />
+    <Compile Include="Hazelcast.Client.Protocol.Codec\RingbufferAddAllCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\RingbufferAddCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\RingbufferCapacityCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\RingbufferHeadSequenceCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\RingbufferMessageType.cs" />
-    <Compile Include="Hazelcast.Client.Protocol.Codec\RingbufferReadManyAsyncCodec.cs" />
+    <Compile Include="Hazelcast.Client.Protocol.Codec\RingbufferReadManyCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\RingbufferReadOneCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\RingbufferRemainingCapacityCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\RingbufferSizeCodec.cs" />
@@ -155,7 +156,6 @@
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapFlushCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapForceUnlockCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapGetAllCodec.cs" />
-    <Compile Include="Hazelcast.Client.Protocol.Codec\MapGetAsyncCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapGetCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapGetEntryViewCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapIsEmptyCodec.cs" />
@@ -165,11 +165,9 @@
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapLockCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapMessageType.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapPutAllCodec.cs" />
-    <Compile Include="Hazelcast.Client.Protocol.Codec\MapPutAsyncCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapPutCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapPutIfAbsentCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapPutTransientCodec.cs" />
-    <Compile Include="Hazelcast.Client.Protocol.Codec\MapRemoveAsyncCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapRemoveCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapRemoveEntryListenerCodec.cs" />
     <Compile Include="Hazelcast.Client.Protocol.Codec\MapRemoveIfSameCodec.cs" />
